+++ Runtime/Utility/LocalizationTemplateResolver.cs
@@ -13,37 +13,37 @@ namespace Localization
         public static string Resolve(string template, LanguageDefinition languageDefinition, LocalizationLookup lookup)
         {
             var placeholders = LocalizationTemplateParser.Parse(template);
-            if(placeholders.Count == 0) return template;
-            
+            if (placeholders.Count == 0) return template;
+
             var subResultDict = new Dictionary<LocalizationPlaceholder, string>();
             foreach (var placeholder in placeholders)
             {
                 if (!lookup.TryGetText(placeholder.NamespaceId,
                         placeholder.Key,
                         languageDefinition,
-                        out var subtemplate)) 
+                        out var subtemplate))
                     continue;
                 var args = placeholder.ArgumentText;
                 var argResults = new string[args.Length];
                 for (int i = 0; i < argResults.Length; i++)
                     argResults[i] = Resolve(args[i], languageDefinition, lookup);
-                
+
                 var subResult = ApplyArguments(subtemplate, argResults);
                 subResultDict[placeholder] = subResult;
             }
-            
+
             var resultBuilder = new StringBuilder(template);
             foreach (var placeholder in placeholders)
             {
-                if(subResultDict.TryGetValue(placeholder, out var subResult))
+                if (subResultDict.TryGetValue(placeholder, out var subResult))
                     resultBuilder.Replace($"<{placeholder.RawContent}>", subResult);
             }
-            
+
             return resultBuilder.ToString();
         }
 
-        private static readonly Regex Regex = new (@"\{\s*(\d+)\s*\}", RegexOptions.Compiled, TimeSpan.FromSeconds(0.5));
-        
+        private static readonly Regex Regex = new(@"\{\s*(\d+)\s*\}", RegexOptions.Compiled, TimeSpan.FromSeconds(0.5));
+
         private static string ApplyArguments(string template, params string[] args)
         {
             return Regex.Replace(template, match =>